窗内 2/3：`op_version` 宽到 u64，run 行的 tag 与清单记同一个数（§9）
**缺陷**：`interface_version` 把 64 位接口哈希截成低 32 位写进清单，而 `interface_tag` 打印
的是补零十六进制的前 8 个字符 —— **高** 32 位。两个面读的是同一个数的两半：`@tag` 从 stdout
抄下来在 `manifest.json` 里**按构造**就找不到（不是碰撞，是分工错了），而两个低 32 位碰撞的
算子会显示同一个 `op_version` 却有不同的键。

**改法**：`ManifestEntry.op_version` → `u64`；`interface_version` 原样返回（它现在是恒等映射，
留着这个名字是为了让"清单里那一栏就是接口哈希"这句话有一个落点）；`interface_tag` 打印完整
16 位。`px_graphs/src/bin/scene.rs` 与 `px_graph/src/driver.rs` 里那两处用 `SHADER_VERSION`
（`u32`）/`SCENE_SCHEMA`（`u32`）的地方显式 `u64::from`：那两个数是**手写的 schema 版本**，
不是接口哈希，改它们的类型是另一件事（反射规则一变要同时 bump 它，见 `docs/invariants.md`）。

**判据**：`cargo test -p px_graph --test keys` 12 passed; 0 failed、
`cargo test -p px_graph_schema` 4 passed、`px_graphs --test inst_gate` 4 passed。
`docs/graph.md` 的清单表同步：那一栏现在是**完整的接口哈希**，不再写"低 32 位"。
⚠ 这次改动落在 `px_graph_schema/src/protocol.rs`（在每份名册里）⇒ 全族节点键轮换，属窗内计划。

+++ px_graph_schema/src/protocol.rs
@@ -9,7 +9,7 @@ pub struct GraphSpec {
 pub struct ManifestEntry {
     pub node: String,
     pub op: String,
-    pub op_version: u32,
+    pub op_version: u64,
     pub key: String,
     pub hit: bool,
     pub millis: u64,
